// src/types.ts
export interface FormData {
    id?: string;
    firstName: string;
    lastName: string;
    phone: string;
    carModel: string;
    licensePlate: string;
    startDateTime: string;
    endDateTime: string;
    parkingArea?: string;
    discountPercentage?: string;
    totalCost?: number;
    currency?: string;
  }
  